requirements.txt
# Library requirements
django-jsonfield
python-dateutil

# Build requirements
setuptools
wheel

# Docs requirements
sphinx
sphinx_rtd_theme

# Test requirements
coverage
tox
codecov
psycopg2-binary
elasticsearch-dsl
elasticsearch
django-grappelli